Add the WebAssembly JS and Web APIs
As discussed in today's WinterCG meeting, although the WebAssembly JS
API (as well as the web API) is built into all mainstream JS engines,
it is not officially part of the JS language. (In fact, Ladybird, the
Serenity OS browser, implements the JS and wasm engines separately.)

As such, since web-interoperable runtimes are expected to support
WebAssembly, it follows that the WebAssembly JS and Web APIs should be
included in the list.